Beta AnswerSheet v0.0.1
The template class for creating the answer sheets has been
completed. Depending on the number of questions, the class arranges
the questions on different pages according to the maximum number of
questions allowed per page. In addition, each page generates its own QR
code where it stores the applicant's information, number of the first
question on the page and number of the last question on the page + 1.